Implement Activity Service: Product Specified Services
This method of implementing the service layer is built around a central registery
that is made accessible to all potential consumers of the services. This allows for
a generic interface to publishing and subscribing, which is both good and bad. It
allows for more flexibility in general, but also places more burden on the implementer
to ensure storage is appropriately tuned for their application.

Pro's
---
- Highly flexible interface
- Allows for optimal storage without requiring additional OEM "re-route" services
- Allows for easier OEM service definitions
- Centralized, unified API for service interfacing

Con's
---
- A bit more obtuse interface; requires stronger Rust foundations to understand
- Requires that tasks are allocated at the product/OEM level rather than in the
  subsystem directly.