type-c-service: Move controller sync state functionality to external API
PD controllers may maintain power and state independently of the EC. If
a port was attached prior to an EC reset then the PD controller might
not trigger an interrupt since nothing has changed on its end. Add a way
to force a sync of the PD controller state to handle this situation.